resume_type 필드를 type으로 변경하고 관련 코드 리팩터링 수행
• Resume 모델의 resume_type 필드를 type으로 변경하여 명확성 향상
• Admin, 매니저, 프록시 모델 등 코드 전반에서 resume_type을 type으로 수정
• 팩토리 코드 재구성 및 중복 제거
• 마이그레이션 파일에서 필드명과 옵션 반영
• 기존 불필요한 마이그레이션 삭제 및 통합
• 구조 개선으로 코드 일관성 및 유지보수성 증대
