+++ webapp/resumes/factories/file_resume_factory.py
@@ -1,0 +1,13 @@
+import factory
+from factory.django import DjangoModelFactory
+from resumes.models import FileResume
+from users.factories import UserFactory
+
+
+class FileResumeFactory(DjangoModelFactory):
+
+  class Meta:
+    model = FileResume
+
+  user = factory.SubFactory(UserFactory)
+  title = factory.Sequence(lambda n: f"파일 이력서 {n}")
