+++ webapp/resumes/factories/resume_factory.py
@@ -1,6 +1,6 @@
 import factory
 from factory.django import DjangoModelFactory
-from resumes.models import FileResume, Resume, TextResume
+from resumes.models import Resume
 from users.factories import UserFactory
 
 
@@ -10,23 +10,5 @@ class ResumeFactory(DjangoModelFactory):
     model = Resume
 
   user = factory.SubFactory(UserFactory)
-  resume_type = "text"
+  type = "text"
   title = factory.Sequence(lambda n: f"이력서 {n}")
-
-
-class TextResumeFactory(DjangoModelFactory):
-
-  class Meta:
-    model = TextResume
-
-  user = factory.SubFactory(UserFactory)
-  title = factory.Sequence(lambda n: f"텍스트 이력서 {n}")
-
-
-class FileResumeFactory(DjangoModelFactory):
-
-  class Meta:
-    model = FileResume
-
-  user = factory.SubFactory(UserFactory)
-  title = factory.Sequence(lambda n: f"파일 이력서 {n}")
