+++ webapp/resumes/models/text_resume.py
@@ -6,11 +6,11 @@ from .resume import Resume
 class TextResumeManager(Resume.objects.__class__):
 
   def get_queryset(self):
-    return super().get_queryset().filter(resume_type=ResumeType.TEXT)
+    return super().get_queryset().filter(type=ResumeType.TEXT)
 
 
 class TextResume(Resume):
-  """resume_type='text' 전용 프록시 모델."""
+  """type='text' 전용 프록시 모델."""
 
   objects = TextResumeManager()
 
@@ -20,5 +20,5 @@ class TextResume(Resume):
     verbose_name_plural = "Text Resumes"
 
   def save(self, *args, **kwargs):
-    self.resume_type = ResumeType.TEXT
+    self.type = ResumeType.TEXT
     super().save(*args, **kwargs)
